0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -12,7 +12,7 @@
 size_t print_list(const list_t *h)
 {
 	const list_t *ap = h;
-	size_t counter = 0;
+	unsigned int  counter = 0;
 
 	while (ap != NULL)
 	{
